[AC97] refactor, split out common code: WIP
